Moved task icons to taskbar folder and tweaked logout style
