+++ src/components/AdminDash/TaskBar/taskIcons.js
@@ -1,0 +1,61 @@
+import React from 'react'
+
+export const PaperMagnify = props => (
+  <svg
+    width='19'
+    height='22'
+    viewBox='0 0 19 22'
+    fill='none'
+    xmlns='http://www.w3.org/2000/svg'
+  >
+    <path
+      d='M6.13037 12.1001C6.13037 12.9753 6.48595 13.8147 7.11889 14.4336C7.75182 15.0524 8.61027 15.4001 9.50537 15.4001C10.4005 15.4001 11.2589 15.0524 11.8919 14.4336C12.5248 13.8147 12.8804 12.9753 12.8804 12.1001C12.8804 11.2249 12.5248 10.3855 11.8919 9.76667C11.2589 9.1478 10.4005 8.80012 9.50537 8.80012C8.61027 8.80012 7.75182 9.1478 7.11889 9.76667C6.48595 10.3855 6.13037 11.2249 6.13037 12.1001ZM18.5054 19.3491V7.86255C18.5054 7.05522 18.18 6.28196 17.6027 5.71753L12.6298 0.855103C12.0692 0.307011 11.3164 0.00012207 10.5324 0.00012207H2.75537C2.15863 0.00012207 1.58634 0.231907 1.16438 0.644487C0.742424 1.05707 0.505371 1.61665 0.505371 2.20012V19.8001C0.505371 20.3836 0.742424 20.9432 1.16438 21.3558C1.58634 21.7683 2.15863 22.0001 2.75537 22.0001H16.2554C16.7616 22.0001 17.2116 21.8351 17.5941 21.5601L12.6104 16.6871C11.7104 17.2591 10.6304 17.6001 9.50537 17.6001C8.01353 17.6001 6.58279 17.0207 5.5279 15.9892C4.473 14.9578 3.88037 13.5588 3.88037 12.1001C3.88037 10.6414 4.473 9.24249 5.5279 8.21103C6.58279 7.17958 8.01353 6.60012 9.50537 6.60012C10.9972 6.60012 12.428 7.17958 13.4828 8.21103C14.5377 9.24249 15.1304 10.6414 15.1304 12.1001C15.1304 13.2001 14.7816 14.2561 14.1966 15.1251L18.5054 19.3491Z'
+      fill={props.color ? props.color : '#fff'}
+    />
+  </svg>
+)
+
+export const Shield = props => (
+  <svg
+    width='19'
+    height='22'
+    viewBox='0 0 19 22'
+    fill='none'
+    xmlns='http://www.w3.org/2000/svg'
+  >
+    <path
+      d='M7.50537 16.0001L3.50537 12.0001L4.91537 10.5901L7.50537 13.1701L14.0954 6.58012L15.5054 8.00012L7.50537 16.0001ZM9.50537 0.00012207L0.505371 4.00012V10.0001C0.505371 15.5501 4.34537 20.7401 9.50537 22.0001C14.6654 20.7401 18.5054 15.5501 18.5054 10.0001V4.00012L9.50537 0.00012207Z'
+      fill={props.color ? props.color : '#fff'}
+    />
+  </svg>
+)
+
+export const PersonCheck = props => (
+  <svg
+    width='27'
+    height='21'
+    viewBox='0 0 27 21'
+    fill='none'
+    xmlns='http://www.w3.org/2000/svg'
+  >
+    <path
+      d='M24.3127 10.8171L26.1794 12.6118L17.4727 21.0001L12.846 16.545L14.7127 14.7503L17.4727 17.3979L24.3127 10.8171ZM9.51272 16.5476L13.5127 20.3662H0.179382V17.8205C0.179382 15.0074 4.95538 12.7289 10.846 12.7289L13.3634 12.8715L9.51272 16.5476ZM10.846 0.00012207C12.2605 0.00012207 13.6171 0.536549 14.6173 1.4914C15.6175 2.44624 16.1794 3.74129 16.1794 5.09165C16.1794 6.44201 15.6175 7.73706 14.6173 8.6919C13.6171 9.64675 12.2605 10.1832 10.846 10.1832C9.43156 10.1832 8.07501 9.64675 7.07481 8.6919C6.07462 7.73706 5.51272 6.44201 5.51272 5.09165C5.51272 3.74129 6.07462 2.44624 7.07481 1.4914C8.07501 0.536549 9.43156 0.00012207 10.846 0.00012207Z'
+      fill={props.color ? props.color : '#fff'}
+    />
+  </svg>
+)
+
+export const HouseMagnify = props => (
+  <svg
+    width='25'
+    height='25'
+    viewBox='0 0 25 25'
+    fill='none'
+    xmlns='http://www.w3.org/2000/svg'
+  >
+    <path
+      d='M20.3801 19.495C20.898 18.649 21.1922 17.6313 21.1922 16.5524C21.1922 13.4871 18.8381 11.0349 15.8955 11.0349C12.9529 11.0349 10.5988 13.4871 10.5988 16.5524C10.5988 19.6176 12.9529 22.0698 15.8955 22.0698C16.9195 22.0698 17.8847 21.7632 18.6969 21.236L22.3693 25.0001L24.0054 23.2959L20.3801 19.495ZM15.8955 19.6176C15.1151 19.6176 14.3666 19.2946 13.8148 18.7198C13.2629 18.145 12.9529 17.3653 12.9529 16.5524C12.9529 15.7394 13.2629 14.9598 13.8148 14.3849C14.3666 13.8101 15.1151 13.4871 15.8955 13.4871C16.6759 13.4871 17.4244 13.8101 17.9763 14.3849C18.5281 14.9598 18.8381 15.7394 18.8381 16.5524C18.8381 17.3653 18.5281 18.145 17.9763 18.7198C17.4244 19.2946 16.6759 19.6176 15.8955 19.6176ZM3.53651 20.8437V11.0349H0.00537109L11.7758 0.00012207L23.5463 11.0349H21.4041C20.0152 9.52685 18.0613 8.58276 15.8955 8.58276C11.6817 8.58276 8.2447 12.1629 8.2447 16.5524C8.2447 18.134 8.69198 19.6176 9.45706 20.8437H3.53651Z'
+      fill={props.color ? props.color : '#fff'}
+    />
+  </svg>
+)
